rearrange callback tests get some more of them out of the old test_integration
