+++ tests/integration/test_integration.py
@@ -1,8 +1,5 @@
 import datetime
 
-from copy import copy
-from multiprocessing import Value
-from selenium.webdriver.common.keys import Keys
 import flask
 
 import pytest
@@ -15,157 +12,10 @@ import dash_flow_example
 import dash_html_components as html
 import dash_core_components as dcc
 
-from dash import Dash, no_update
+from dash import Dash
 
-from dash.dependencies import Input, Output, State
-from dash.exceptions import (
-    PreventUpdate,
-    InvalidCallbackReturnValue,
-    IncorrectTypeException,
-)
-from dash.testing.wait import until
-
-
-def test_inin001_simple_callback(dash_duo):
-    app = Dash(__name__)
-    app.layout = html.Div(
-        [
-            dcc.Input(id="input", value="initial value"),
-            html.Div(html.Div([1.5, None, "string", html.Div(id="output-1")])),
-        ]
-    )
-
-    call_count = Value("i", 0)
-
-    @app.callback(Output("output-1", "children"), [Input("input", "value")])
-    def update_output(value):
-        call_count.value += 1
-        return value
-
-    dash_duo.start_server(app)
-
-    dash_duo.wait_for_text_to_equal("#output-1", "initial value")
-    dash_duo.percy_snapshot(name="simple-callback-1")
-
-    input1 = dash_duo.find_element("#input")
-    dash_duo.clear_input(input1)
-    input1.send_keys("hello world")
-
-    dash_duo.wait_for_text_to_equal("#output-1", "hello world")
-    dash_duo.percy_snapshot(name="simple-callback-2")
-
-    # an initial call, one for clearing the input
-    # and one for each hello world character
-    assert call_count.value == 2 + len("hello world")
-
-    assert not dash_duo.get_logs()
-
-
-def test_inin002_wildcard_callback(dash_duo):
-    app = Dash(__name__)
-    app.layout = html.Div(
-        [
-            dcc.Input(id="input", value="initial value"),
-            html.Div(
-                html.Div(
-                    [
-                        1.5,
-                        None,
-                        "string",
-                        html.Div(
-                            id="output-1",
-                            **{"data-cb": "initial value", "aria-cb": "initial value"}
-                        ),
-                    ]
-                )
-            ),
-        ]
-    )
-
-    input_call_count = Value("i", 0)
-
-    @app.callback(Output("output-1", "data-cb"), [Input("input", "value")])
-    def update_data(value):
-        input_call_count.value += 1
-        return value
-
-    @app.callback(Output("output-1", "children"), [Input("output-1", "data-cb")])
-    def update_text(data):
-        return data
-
-    dash_duo.start_server(app)
-    dash_duo.wait_for_text_to_equal("#output-1", "initial value")
-    dash_duo.percy_snapshot(name="wildcard-callback-1")
-
-    input1 = dash_duo.find_element("#input")
-    dash_duo.clear_input(input1)
-    input1.send_keys("hello world")
-
-    dash_duo.wait_for_text_to_equal("#output-1", "hello world")
-    dash_duo.percy_snapshot(name="wildcard-callback-2")
-
-    # an initial call, one for clearing the input
-    # and one for each hello world character
-    assert input_call_count.value == 2 + len("hello world")
-
-    assert not dash_duo.get_logs()
-
-
-def test_inin003_aborted_callback(dash_duo):
-    """Raising PreventUpdate OR returning no_update prevents update and
-    triggering dependencies."""
-
-    initial_input = "initial input"
-    initial_output = "initial output"
-
-    app = Dash(__name__)
-    app.layout = html.Div(
-        [
-            dcc.Input(id="input", value=initial_input),
-            html.Div(initial_output, id="output1"),
-            html.Div(initial_output, id="output2"),
-        ]
-    )
-
-    callback1_count = Value("i", 0)
-    callback2_count = Value("i", 0)
-
-    @app.callback(Output("output1", "children"), [Input("input", "value")])
-    def callback1(value):
-        callback1_count.value += 1
-        if callback1_count.value > 2:
-            return no_update
-        raise PreventUpdate("testing callback does not update")
-        return value
-
-    @app.callback(Output("output2", "children"), [Input("output1", "children")])
-    def callback2(value):
-        callback2_count.value += 1
-        return value
-
-    dash_duo.start_server(app)
-
-    input_ = dash_duo.find_element("#input")
-    input_.send_keys("xyz")
-    dash_duo.wait_for_text_to_equal("#input", "initial inputxyz")
-
-    until(
-        lambda: callback1_count.value == 4,
-        timeout=3,
-        msg="callback1 runs 4x (initial page load and 3x through send_keys)",
-    )
-
-    assert (
-        callback2_count.value == 0
-    ), "callback2 is never triggered, even on initial load"
-
-    # double check that output1 and output2 children were not updated
-    assert dash_duo.find_element("#output1").text == initial_output
-    assert dash_duo.find_element("#output2").text == initial_output
-
-    assert not dash_duo.get_logs()
-
-    dash_duo.percy_snapshot(name="aborted")
+from dash.dependencies import Input, Output
+from dash.exceptions import PreventUpdate
 
 
 def test_inin004_wildcard_data_attributes(dash_duo):
@@ -372,104 +222,6 @@ def test_inin010_func_layout_accepted(dash_duo):
     assert dash_duo.find_element("#a").text == "Hello World"
 
 
-def test_inin012_multi_output_no_update(dash_duo):
-    app = Dash(__name__)
-
-    app.layout = html.Div(
-        [
-            html.Button("B", "btn"),
-            html.P("initial1", "n1"),
-            html.P("initial2", "n2"),
-            html.P("initial3", "n3"),
-        ]
-    )
-
-    @app.callback(
-        [Output("n1", "children"), Output("n2", "children"), Output("n3", "children")],
-        [Input("btn", "n_clicks")],
-    )
-    def show_clicks(n):
-        # partial or complete cancelation of updates via no_update
-        return [
-            no_update if n and n > 4 else n,
-            no_update if n and n > 2 else n,
-            # make a new instance, to mock up caching and restoring no_update
-            copy(no_update),
-        ]
-
-    dash_duo.start_server(app)
-
-    dash_duo.multiple_click("#btn", 10)
-
-    dash_duo.wait_for_text_to_equal("#n1", "4")
-    dash_duo.wait_for_text_to_equal("#n2", "2")
-    dash_duo.wait_for_text_to_equal("#n3", "initial3")
-
-
-def test_inin013_no_update_chains(dash_duo):
-    app = Dash(__name__)
-
-    app.layout = html.Div(
-        [
-            dcc.Input(id="a_in", value="a"),
-            dcc.Input(id="b_in", value="b"),
-            html.P("", id="a_out"),
-            html.P("", id="a_out_short"),
-            html.P("", id="b_out"),
-            html.P("", id="ab_out"),
-        ]
-    )
-
-    @app.callback(
-        [Output("a_out", "children"), Output("a_out_short", "children")],
-        [Input("a_in", "value")],
-    )
-    def a_out(a):
-        return a, a if len(a) < 3 else no_update
-
-    @app.callback(Output("b_out", "children"), [Input("b_in", "value")])
-    def b_out(b):
-        return b
-
-    @app.callback(
-        Output("ab_out", "children"),
-        [Input("a_out_short", "children")],
-        [State("b_out", "children")],
-    )
-    def ab_out(a, b):
-        return a + " " + b
-
-    dash_duo.start_server(app)
-
-    a_in = dash_duo.find_element("#a_in")
-    b_in = dash_duo.find_element("#b_in")
-
-    b_in.send_keys("b")
-    a_in.send_keys("a")
-    dash_duo.wait_for_text_to_equal("#a_out", "aa")
-    dash_duo.wait_for_text_to_equal("#b_out", "bb")
-    dash_duo.wait_for_text_to_equal("#a_out_short", "aa")
-    dash_duo.wait_for_text_to_equal("#ab_out", "aa bb")
-
-    b_in.send_keys("b")
-    a_in.send_keys("a")
-    dash_duo.wait_for_text_to_equal("#a_out", "aaa")
-    dash_duo.wait_for_text_to_equal("#b_out", "bbb")
-    dash_duo.wait_for_text_to_equal("#a_out_short", "aa")
-    # ab_out has not been triggered because a_out_short received no_update
-    dash_duo.wait_for_text_to_equal("#ab_out", "aa bb")
-
-    b_in.send_keys("b")
-    a_in.send_keys(Keys.END)
-    a_in.send_keys(Keys.BACKSPACE)
-    dash_duo.wait_for_text_to_equal("#a_out", "aa")
-    dash_duo.wait_for_text_to_equal("#b_out", "bbbb")
-    dash_duo.wait_for_text_to_equal("#a_out_short", "aa")
-    # now ab_out *is* triggered - a_out_short got a new value
-    # even though that value is the same as the last value it got
-    dash_duo.wait_for_text_to_equal("#ab_out", "aa bbbb")
-
-
 def test_inin014_with_custom_renderer(dash_duo):
     app = Dash(__name__)
 
@@ -642,110 +394,6 @@ def test_inin017_late_component_register(dash_duo):
     btn.click()
 
     dash_duo.find_element("#inserted-input")
-
-
-def test_inin019_callback_dep_types():
-    app = Dash(__name__)
-    app.layout = html.Div(
-        [
-            html.Div("child", id="in1"),
-            html.Div("state", id="state1"),
-            html.Div(id="out1"),
-            html.Div("child", id="in2"),
-            html.Div("state", id="state2"),
-            html.Div(id="out2"),
-            html.Div("child", id="in3"),
-            html.Div("state", id="state3"),
-            html.Div(id="out3"),
-        ]
-    )
-
-    with pytest.raises(IncorrectTypeException):
-
-        @app.callback([[Output("out1", "children")]], [Input("in1", "children")])
-        def f(i):
-            return i
-
-        pytest.fail("extra output nesting")
-
-    # all OK with tuples
-    @app.callback(
-        (Output("out1", "children"),),
-        (Input("in1", "children"),),
-        (State("state1", "children"),),
-    )
-    def f1(i):
-        return i
-
-    # all OK with all args in single list
-    @app.callback(
-        Output("out2", "children"),
-        Input("in2", "children"),
-        State("state2", "children"),
-    )
-    def f2(i):
-        return i
-
-    # all OK with lists
-    @app.callback(
-        [Output("out3", "children")],
-        [Input("in3", "children")],
-        [State("state3", "children")],
-    )
-    def f3(i):
-        return i
-
-
-def test_inin020_callback_return_validation():
-    app = Dash(__name__)
-    app.layout = html.Div(
-        [
-            html.Div(id="a"),
-            html.Div(id="b"),
-            html.Div(id="c"),
-            html.Div(id="d"),
-            html.Div(id="e"),
-            html.Div(id="f"),
-        ]
-    )
-
-    @app.callback(Output("b", "children"), [Input("a", "children")])
-    def single(a):
-        return set([1])
-
-    with pytest.raises(InvalidCallbackReturnValue):
-        # outputs_list (normally callback_context.outputs_list) is provided
-        # by the dispatcher from the request.
-        single("aaa", outputs_list={"id": "b", "property": "children"})
-        pytest.fail("not serializable")
-
-    @app.callback(
-        [Output("c", "children"), Output("d", "children")], [Input("a", "children")]
-    )
-    def multi(a):
-        return [1, set([2])]
-
-    with pytest.raises(InvalidCallbackReturnValue):
-        outputs_list = [
-            {"id": "c", "property": "children"},
-            {"id": "d", "property": "children"},
-        ]
-        multi("aaa", outputs_list=outputs_list)
-        pytest.fail("nested non-serializable")
-
-    @app.callback(
-        [Output("e", "children"), Output("f", "children")], [Input("a", "children")]
-    )
-    def multi2(a):
-        return ["abc"]
-
-    with pytest.raises(InvalidCallbackReturnValue):
-        outputs_list = [
-            {"id": "e", "property": "children"},
-            {"id": "f", "property": "children"},
-        ]
-        multi2("aaa", outputs_list=outputs_list)
-        pytest.fail("wrong-length list")
 
 
 def test_inin_024_port_env_success(dash_duo):
